Fix: Settlement Name, ReceiptItem 수정하기
ReceiptItem 수정 시 숫자만 입력가능,
모두 지워지면 0으로 초기화,
Settlement Name Appbar위치로 옮김
