+++ lib/ViewModel/mainviewmodel.dart
@@ -13,14 +13,15 @@ class MainViewModel extends ChangeNotifier {
   Settlement selectedSettlement = Settlement();
   List<List<List<TextEditingController>>> receiptItemControllerList = [];
 
-  void updateTotalPrice(receiptIndex){
+  void updateTotalPrice(receiptIndex) {
     double total = 0;
-    for(ReceiptItem receiptItem in selectedSettlement.receipts[receiptIndex].receiptItems){
+    for (ReceiptItem receiptItem
+        in selectedSettlement.receipts[receiptIndex].receiptItems) {
       total += receiptItem.price;
     }
     selectedSettlement.receipts[receiptIndex].totalPrice = total;
     total = 0;
-    for(Receipt receipt in selectedSettlement.receipts){
+    for (Receipt receipt in selectedSettlement.receipts) {
       total += receipt.totalPrice;
     }
     selectedSettlement.totalPrice = total;
@@ -54,9 +55,15 @@ class MainViewModel extends ChangeNotifier {
         newIndividualPrice *
             selectedSettlement
                 .receipts[receiptIndex].receiptItems[receiptItemIndex].count;
+    receiptItemControllerList[receiptIndex][receiptItemIndex][1].text =
+        selectedSettlement.receipts[receiptIndex].receiptItems[receiptItemIndex]
+            .individualPrice
+            .truncate()
+            .toString();
     receiptItemControllerList[receiptIndex][receiptItemIndex][3].text =
         selectedSettlement
-            .receipts[receiptIndex].receiptItems[receiptItemIndex].price.truncate()
+            .receipts[receiptIndex].receiptItems[receiptItemIndex].price
+            .truncate()
             .toString();
     updateTotalPrice(receiptIndex);
     notifyListeners();
@@ -73,9 +80,10 @@ class MainViewModel extends ChangeNotifier {
                 .receiptItems[receiptItemIndex].individualPrice;
     receiptItemControllerList[receiptIndex][receiptItemIndex][3].text =
         selectedSettlement
-            .receipts[receiptIndex].receiptItems[receiptItemIndex].price.truncate()
+            .receipts[receiptIndex].receiptItems[receiptItemIndex].price
+            .truncate()
             .toString();
-    
+
     updateTotalPrice(receiptIndex);
     notifyListeners();
   }
@@ -91,9 +99,15 @@ class MainViewModel extends ChangeNotifier {
                 .receipts[receiptIndex].receiptItems[receiptItemIndex].count;
     receiptItemControllerList[receiptIndex][receiptItemIndex][1].text =
         selectedSettlement.receipts[receiptIndex].receiptItems[receiptItemIndex]
-            .individualPrice.truncate()
+            .individualPrice
+            .truncate()
             .toString();
-    
+    receiptItemControllerList[receiptIndex][receiptItemIndex][3].text =
+        selectedSettlement
+            .receipts[receiptIndex].receiptItems[receiptItemIndex].price
+            .truncate()
+            .toString();
+
     updateTotalPrice(receiptIndex);
     notifyListeners();
   }
